+++ priv/static/phoenix_kit_reload.js
@@ -1,0 +1,1 @@
+// Auto-reload trigger 1786800105043049460